Laboratorium4/requirements.txt
Django==3.1.2
djangorestframework==3.12.2
drf-yasg==1.20.0
django-filter